+++ scripts/sign.js
