+++ scripts/signup.js
